backend/requirements.txt
# Flask Backend Dependencies
# Python 3.10.11 compatible versions

# Flask Core
Flask==3.0.0
Flask-CORS==4.0.0

# Database
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
PyMySQL==1.1.0
cryptography==41.0.7

# Authentication
Flask-JWT-Extended==4.6.0
bcrypt==4.1.2
google-auth==2.23.0

# Serialization & Validation
marshmallow==3.20.1
Flask-Marshmallow==0.15.0
marshmallow-sqlalchemy==0.29.0

# File Upload
cloudinary==1.40.0

# Environment
python-dotenv==1.0.0

# Production Server
gunicorn==21.2.0

# Development (optional)
# Flask-DebugToolbar==0.14.1
